Moved Compartment and Segment struct into separate classes
