+++ FieldOpt/Model/wells/segmented_well.h
@@ -22,8 +22,10 @@
 #define FIELDOPT_SEGMENTED_WELL_H
 
 #include "well.h"
+#include "segment.h"
 #include "Model/wells/wellbore/completions/packer.h"
 #include "Model/wells/wellbore/completions/icd.h"
+#include "compartment.h"
 
 namespace Model {
 namespace Wells {
@@ -44,68 +46,6 @@ class SegmentedWell : public Well {
                 Properties::VariablePropertyContainer *variable_container,
                 Reservoir::Grid::Grid *grid);
 
-  /*!
-   * The Compartment struct models a compartment made up of two packers and an ICD.
-   * The ICD is placed at the beginning of the compartment (i.e. closest to the
-   * heel of the well).
-   */
-  struct Compartment {
-    Compartment(const double start_md, const double end_md,
-                const double start_tvd, const double end_tvd,
-                const Settings::Model::Well &well_settings,
-                Properties::VariablePropertyContainer *variable_container,
-                std::vector<Compartment> &compartments_);
-    double GetLength() const;
-    double GetTVDDifference() const;
-    Packer *start_packer;
-    Packer *end_packer;
-    ICD *icd;
-  };
-
-  /*!
-   * The Segment struct contains the data needed by the simulator interface
-   * to build the WELSEGS, WSEGVALV and COMPSEGS keywords.
-   *
-   * Annulus segments should only pass through _one_ well block.
-   */
-  struct Segment {
-    enum SegType { TUBING_SEGMENT, ANNULUS_SEGMENT, ICD_SEGMENT };
-    Segment();
-    Segment(SegType type, int index, int branch, int outlet,
-            double length, double tvd, double diameter, double roughness,
-            double outlet_md
-    );
-    void AddInlet(int index);
-    std::vector<int> GetInlets() const;
-    void AddParentBlock(Wellbore::WellBlock *parent_block);
-    void AddParentICD(Wellbore::Completions::ICD *parent_icd);
-    bool    HasParentBlock() const; //!< Check whether this segment has a parent well block.
-    bool    HasParentICD()   const;   //!< Check whether this segment has a parent ICD decvice.
-    Wellbore::WellBlock *ParentBlock() { return parent_block_; }
-    Wellbore::Completions::ICD *ParentICD() { return parent_icd_; }
-    int     Index()     const { return index_; }
-    int     Branch()    const { return branch_; }
-    int     Outlet()    const { return outlet_; }
-    int     OutletMD()  const { return md_; }
-    double  TVDChange() const { return tvd_change_; }
-    double  Length()    const { return length_; }
-    double  Diameter()  const { return diameter_; }
-    double  Roughness() const { return roughness_; }
-    SegType Type()      const { return type_; }
-    std::string ToString();
-
-   private:
-    int index_;                            //!< Segment index
-    int branch_;                           //!< Branch index
-    int outlet_;                           //!< Index of outlet segment.
-    double length_, diameter_, roughness_; //!< Physical segment properties
-    double md_;                            //!< Measured depth to start of segment
-    double tvd_change_;                    //!< TVD _change_ along the segment (i.e. zero when perfectly horizontal)
-    SegType type_;                         //!< Segment type (tubing, annulus or ICD)
-    std::vector<int> inlets_;              //!< List of segment (indices) with this segment as their outlet.
-    Wellbore::WellBlock *parent_block_ = 0;      //!< Parent well block for annulus segments.
-    Wellbore::Completions::ICD *parent_icd_ = 0; //!< Parent ICD for ICD segments.
-  };
 
   std::vector<Compartment> GetCompartments() const;
   std::vector<Packer *> GetPackers() const;
@@ -114,6 +54,7 @@ class SegmentedWell : public Well {
   std::vector<Segment> GetTubingSegments();
   std::vector<Segment> GetICDSegments();
   std::vector<Segment> GetAnnulusSegments();
+  virtual bool IsSegmented() const { return true; }
 
 
  private:
